feat(tests): mv feature_to_block tests to test_block_to_feature.py
